requirements.txt
# Core Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Vision & ML
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
timm==0.9.12
pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0
numpy==1.24.3
scipy==1.11.4

# Image Processing & Similarity
imagehash==4.3.1

# Database & Search
duckdb==0.9.2
faiss-cpu==1.7.4

# Web Scraping & Monitoring
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
selenium==4.15.0
lxml==4.9.3

# Report Generation
reportlab==4.0.7
python-docx==1.1.0

# Configuration & Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0

# Utilities
pandas==2.1.3
tqdm==4.66.1
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Optional: LangGraph for agent orchestration (can be added later)
# langgraph==0.0.20
# langchain==0.1.0
